fix: Change stageProgress from Int to Double to fix JSON parsing crash
The API returns stageProgress as a floating-point number (e.g., 0.33495145631068),
but it was declared as Int which caused JSON deserialization to fail when
encountering the decimal point.

Changes:
- ReadaloudApiModel: stageProgress Int? -> Double?
- ReadaloudDomainModel: stageProgress Int? -> Double?
- ReadaloudLocalModel: stageProgress Int? -> Double?
- ReadaloudEntity: stageProgress Int? -> Double?
- ReadaloudSqlDelightEntity: stageProgress Int? -> Double?
- Readaloud.sq: stage_progress INTEGER -> REAL
- BooksSqlDelightDao: Remove toLong/toInt conversions for stageProgress
- Bump database schema version from 2 to 3
- Add migration file 2.sqm

+++ feature/books/data/src/commonMain/kotlin/com/retro99/books/data/model/ReadaloudApiModel.kt
@@ -22,7 +22,7 @@ data class ReadaloudApiModel(
     val currentStage: String? = null,
 
     @SerialName("stageProgress")
-    val stageProgress: Int? = null,
+    val stageProgress: Double? = null,
 
     @SerialName("queuePosition")
     val queuePosition: Int? = null,
